# schemas/openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: RaptorFlow API
  version: 0.1.0
servers:
  - url: https://api.raptorflow.example.com
paths:
...
  /api/v1/council:
    post:
      summary: Start council session
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                campaign_id:
                  type: string
                agent_roster:
                  type: array
                  items:
                    type: string
                question:
                  type: string
                session_type:
                  type: string
      responses:
        "201":
          description: Session created
        "400":
          description: Invalid request
    get:
      summary: List council sessions
      responses:
        "200":
          description: Session list
  /api/v1/council/{session_id}:
    get:
      summary: Get council session with positions
      responses:
        "200":
          description: Session detail
        "404":
          description: Session not found
  /api/v1/council/{session_id}/messages:
    get:
      summary: Get agent positions for a session
      responses:
        "200":
          description: Agent positions
        "404":
          description: Session not found
  /api/v1/muse:
    post:
      summary: Submit prompt to Muse
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                route:
                  type: string
                conversation_id:
                  type: string
      responses:
        "200":
          description: Prompt submitted
        "400":
          description: Prompt required
    get:
      summary: List Muse conversations
      responses:
        "200":
          description: Conversation list
  /api/v1/muse/{conversation_id}:
    get:
      summary: Get Muse conversation with messages
      responses:
        "200":
          description: Conversation detail
        "404":
          description: Conversation not found
  /api/v1/muse/{conversation_id}/messages:
    get:
      summary: Get messages for a Muse conversation
      responses:
        "200":
          description: Message list
        "404":
          description: Conversation not found
...
  /api/v1/daily-wins:
    get:
      summary: List daily wins
      responses:
        "200":
          description: Daily win list
    post:
      summary: Create daily win
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                [
                  briefing_date,
                  lead_summary,
                  full_briefing,
                  recommended_action,
                  recommended_action_type,
                ]
              properties:
                briefing_date:
                  type: string
                  format: date
                lead_summary:
                  type: string
                full_briefing:
                  type: string
                recommended_action:
                  type: string
                recommended_action_type:
                  type: string
                recommended_action_data:
                  type: object
      responses:
        "201":
          description: Daily win created
        "400":
          description: Invalid request
...
  /api/v1/daily-wins/{id}/viewed:
    post:
      summary: Mark daily win as viewed
      responses:
        "200":
          description: Status updated
        "404":
          description: Daily win not found
  /api/v1/daily-wins/{id}/acted:
    post:
      summary: Mark daily win as acted upon
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome:
                  type: string
      responses:
        "200":
          description: Status updated
        "404":
          description: Daily win not found
...